+++ Entities/Helper/Config.php
@@ -7,6 +7,16 @@ use \Magento\Framework\App\Helper\AbstractHelper;
 class Config extends AbstractHelper
 {
     /**
+     * Data in file insertion method
+     */
+    const INSERTION_METHOD_DATA_IN_FILE = 'data_in_file';
+
+    /**
+     * By rows insertion method
+     */
+    const INSERTION_METHOD_BY_ROWS = 'by_rows';
+
+    /**
      * Retrieve CSV configuration
      *
      * @return array
@@ -28,6 +38,16 @@ class Config extends AbstractHelper
     public function getLoadDataLocal()
     {
         return $this->scopeConfig->getValue('pimgento/general/load_data_local');
+    }
+
+    /**
+     * Retrieve insertion method
+     *
+     * @return string
+     */
+    public function getInsertionMethod()
+    {
+        return (string) $this->scopeConfig->getValue('pimgento/general/data_insertion_method');
     }
 
 }